+++ src/components/grid/columnTemplate.context.ts
@@ -1,9 +1,9 @@
-import {ColumnComponent} from './column.component';
+import {ColumnLegacyComponent} from './column.component';
 
 /**
  * Context for column template
  */
-export class ColumnTemplateContext
+export class ColumnTemplateLegacyContext
 {
     //######################### public properties #########################
 
@@ -21,12 +21,12 @@ export class ColumnTemplateContext
     /**
      * Creates instance of ColumnTemplateContext
      * @param  {any} $implicit Data of current row
-     * @param  {ColumnComponent} column Object of column itself
+     * @param  {ColumnLegacyComponent} column Object of column itself
      * @param  {number} index Index of rendered row in current page
      * @param  {number[]} rowIndexes Indexes of rows for whole data mapped for current page
      */
     constructor(public $implicit: any, 
-                public column: ColumnComponent, 
+                public column: ColumnLegacyComponent, 
                 public index: number, 
                 public rowIndexes: number[])
     {