+++ src/components/grid/gridCookie.config.ts
@@ -1,7 +1,7 @@
 /**
  *  Object that is stored in cookie for each grid
  */
-export interface GridCookieConfig
+export interface GridCookieLegacyConfig
 {
     /**
      * Array of ordered booleans indicating visibility of columns